Properly test type constants, and add one for GStrv.
The constants are defined using type_from_name, so use type_name to close
the circle in the test, rather than using the same code again.

The GStrv type is currently not loaded by the time these constants are
defined, so it is necessary to use GLib.strv_get_type to get its GType.