Make enum a class
The enum builder returns a class, descended from EnumBase. Within the
class, the constant Enum is defined, whose value is the original
FFI::Enum.

To make attach_function work in the library modules, this constant needs
to be passed. Therefore, IEnumInfo.to_ffitype is changed to return that
constant.

Several helper functions are updated to handle the changed nature of
enums.

Still to do:

- Change from a class to a module.
- Delegate more behavior to the underlying enum.
- Define EnumBase.from to bring its behavior in line with the other
  types. This will simplify the argument builder code.

- Down the road, something similar will need to happen to callback
  functions.