Add callback exception tests
Demonstrates the problem described in issue #40. The problem occurs when
The exception is raised inside callbacks executed in an event loop.
These exceptions are swallowed by the system. The desired behavior is
for them to abort the event loop, causing the MainLoop#run method to
return with a raised exception.