Implement assignment to struct fields
Makes struct classes descend from a new base class, StructBase, which is
a DataConverter. Due to this, struct classes are now their own FFI type.
This means they can be used as-is in a layout specification, and no
longer cause a type mismatch during assignment.

The test was adjusted to take into account that this is a value
assignment, not the assignment of a pointer.