Handle truly callee-allocated simple types
By default, an out argument will always be marked as caller-allocated=0.
If the C type is a pointer to a simple type, this makes no sense. For
example, a gint* needs to be allocated by the caller, or else the caller
will not know where the value will exist. However, a gint** can be
allocated by the callee (at least the inner pointer). This commit adds
support for the latter case for both functions and callbacks.