Give handler loop test a little more time to run
This makes sure the main_loop test finishes properly on Ruby 1.9.3. The
time given should allow the handler to run three times.

What should normally happen is this:
- The first run of the handler makes the thread start. It immediately
  goes to sleep.
- The second run wakes up the thread which inserts an element into the
  array.

However, in some cases it is only at the second run of the handler that
the thread is first started. In that case the insertion of the element
takes place too late.