Work-around for field building for GSignalQuery
GSignalQuery's introspection data uses the array_length info to link the
param_types field and the n_params field. GirFFI's field builder is
confused by this, because array_length doesn't point to an argument but
to a field. Work around this by ignore this info if the corresponding
argument cannot be found.