Make .from_enumerable methods for lists work with all enumerables
The #reverse method is not defined in Enumerable so should not be used.
Instead, #reverse_each could have been used, but it creates a temporary
array so is potentially slow. It also causes an infinite loop for
infinite ranges, but that is not really relevant since #reduce also
causes an infinite loop.