Handle functions that have swapped closure annotation
In nearly all cases, if a function has a callable parameter and a user
data parameter, the closure annotation points from the callable to the
user data. However, from gobject-introspection 1.67.1, the
g_signal_handlers_unblock_matched function is annotated the other way
around.

This change checks for this and takes appropriate action.