Inline slightly broken #set_property= and #get_property= methods
These methods only worked on objects of the class GObject::ObjectClass
itself, but not on derived classes, since those would have a struct
lacking the :set_property and :get_property fields.

Since these methods are only used internally and in one place, inline
them so we know they will work with the right kind of struct.